+++ phonenumber_field/formfields.py
@@ -1,24 +1,39 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
+import phonenumbers
 from django.core import validators
 from django.core.exceptions import ValidationError
 from django.forms.fields import CharField
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext as _
 
 from phonenumber_field.phonenumber import to_python, validate_region
 from phonenumber_field.validators import validate_international_phonenumber
 
 
 class PhoneNumberField(CharField):
-    default_error_messages = {"invalid": _("Enter a valid phone number.")}
     default_validators = [validate_international_phonenumber]
 
-    def __init__(self, *args, **kwargs):
-        self.region = kwargs.pop("region", None)
-        validate_region(self.region)
-        super(PhoneNumberField, self).__init__(*args, **kwargs)
+    def __init__(self, *args, region=None, **kwargs):
+        super().__init__(*args, **kwargs)
         self.widget.input_type = "tel"
+
+        validate_region(region)
+        self.region = region
+
+        if "invalid" not in self.error_messages:
+            if region:
+                number = phonenumbers.example_number(region)
+                example_number = to_python(number).as_national
+                # Translators: {example_number} is a national phone number.
+                error_message = _(
+                    "Enter a valid phone number (e.g. {example_number}) "
+                    "or a number with an international call prefix."
+                )
+            else:
+                example_number = "+12125552368"  # Ghostbusters
+                # Translators: {example_number} is an international phone number.
+                error_message = _("Enter a valid phone number (e.g. {example_number}).")
+            self.error_messages["invalid"] = error_message.format(
+                example_number=example_number
+            )
 
     def to_python(self, value):
         phone_number = to_python(value, region=self.region)