+++ tests/tests.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 import phonenumbers
 from django import forms
 from django.core import checks
@@ -372,6 +369,39 @@ class PhonenumerFieldAppTest(TestCase):
         self.assertIsInstance(models.TestModel.phone, modelfields.PhoneNumberDescriptor)
 
 
+class PhoneNumberFormFieldTest(TestCase):
+    def test_error_message(self):
+        class PhoneNumberForm(forms.Form):
+            number = formfields.PhoneNumberField()
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(
+            form.errors, {"number": ["Enter a valid phone number (e.g. +12125552368)."]}
+        )
+
+    def test_override_error_message(self):
+        class MyPhoneNumberField(formfields.PhoneNumberField):
+            default_error_messages = {"invalid": "MY INVALID MESSAGE!"}
+
+        class PhoneNumberForm(forms.Form):
+            number = MyPhoneNumberField()
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(form.errors, {"number": ["MY INVALID MESSAGE!"]})
+
+    def test_override_error_message_inline(self):
+        class PhoneNumberForm(forms.Form):
+            number = formfields.PhoneNumberField(
+                error_messages={"invalid": "MY INLINE INVALID MESSAGE!"}
+            )
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(form.errors, {"number": ["MY INLINE INVALID MESSAGE!"]})
+
+
 class RegionPhoneNumberFormFieldTest(TestCase):
     def test_regional_phone(self):
         class PhoneNumberForm(forms.Form):
@@ -392,6 +422,22 @@ class RegionPhoneNumberFormFieldTest(TestCase):
 
         self.assertTrue(
             force_text(cm.exception).startswith("“invalid” is not a valid region code.")
+        )
+
+    def test_error_message_nationalize_example(self):
+        class PhoneNumberForm(forms.Form):
+            number = formfields.PhoneNumberField(region="CA")
+
+        form = PhoneNumberForm({"number": "invalid"})
+        self.assertIs(form.is_valid(), False)
+        self.assertEqual(
+            form.errors,
+            {
+                "number": [
+                    "Enter a valid phone number (e.g. (506) 234-5678) "
+                    "or a number with an international call prefix."
+                ]
+            },
         )